requirements.txt
gradio
requests
python-dotenv
pydub
TTS
